refactor TestHandleAlreadyJailed to simapp
