+++ x/slashing/keeper/keeper_test.go
@@ -63,3 +63,57 @@ func TestHandleNewValidator(t *testing.T) {
 	expTokens := sdk.TokensFromConsensusPower(100)
 	require.Equal(t, expTokens.Int64(), app.BankKeeper.GetBalance(ctx, bondPool.GetAddress(), app.StakingKeeper.BondDenom(ctx)).Amount.Int64())
 }
+
+// Test a jailed validator being "down" twice
+// Ensure that they're only slashed once
+func TestHandleAlreadyJailed(t *testing.T) {
+	// initial setup
+	app := simapp.Setup(false)
+	ctx := app.BaseApp.NewContext(false, abci.Header{})
+	power := int64(100)
+
+	amt := sdk.TokensFromConsensusPower(power)
+	addrDels := simapp.AddTestAddrsIncremental(app, ctx, 1, sdk.TokensFromConsensusPower(200))
+	valAddrs := simapp.ConvertAddrsToValAddrs(addrDels)
+	pks := simapp.CreateTestPubKeys(1)
+
+	addr, val := valAddrs[0], pks[0]
+	sh := staking.NewHandler(app.StakingKeeper)
+	res, err := sh(ctx, keeper.NewTestMsgCreateValidator(addr, val, amt))
+	require.NoError(t, err)
+	require.NotNil(t, res)
+
+	staking.EndBlocker(ctx, app.StakingKeeper)
+
+	// 1000 first blocks OK
+	height := int64(0)
+	for ; height < app.SlashingKeeper.SignedBlocksWindow(ctx); height++ {
+		ctx = ctx.WithBlockHeight(height)
+		app.SlashingKeeper.HandleValidatorSignature(ctx, val.Address(), power, true)
+	}
+
+	// 501 blocks missed
+	for ; height < app.SlashingKeeper.SignedBlocksWindow(ctx)+(app.SlashingKeeper.SignedBlocksWindow(ctx)-app.SlashingKeeper.MinSignedPerWindow(ctx))+1; height++ {
+		ctx = ctx.WithBlockHeight(height)
+		app.SlashingKeeper.HandleValidatorSignature(ctx, val.Address(), power, false)
+	}
+
+	// end block
+	staking.EndBlocker(ctx, app.StakingKeeper)
+
+	// validator should have been jailed and slashed
+	validator, _ := app.StakingKeeper.GetValidatorByConsAddr(ctx, sdk.GetConsAddress(val))
+	require.Equal(t, sdk.Unbonding, validator.GetStatus())
+
+	// validator should have been slashed
+	resultingTokens := amt.Sub(sdk.TokensFromConsensusPower(1))
+	require.Equal(t, resultingTokens, validator.GetTokens())
+
+	// another block missed
+	ctx = ctx.WithBlockHeight(height)
+	app.SlashingKeeper.HandleValidatorSignature(ctx, val.Address(), power, false)
+
+	// validator should not have been slashed twice
+	validator, _ = app.StakingKeeper.GetValidatorByConsAddr(ctx, sdk.GetConsAddress(val))
+	require.Equal(t, resultingTokens, validator.GetTokens())
+}
